Utilities

These are a random collection of functions to do with counting sizes, record sizes, and dictionary manipulation functions.

Note

They are just here as this is a preliminary version of the package.

Kmer dictionary generation

KmerGMA.genKmers — Function

genKmers(k::Int64,
         Dictionary::Bool = true;
         withN::Bool = false,
         cumulative::Bool = false)

systematic Kmer generation into dictionary：It matches each kmer to a unique index.

the withN argument can be specified to decide whether dna"N" should be included as a nucleotide in the resulting kmers.

the cumulative argument shouldnt ever be needed but it generates a kmer Dictionary with all the kmer sizes up to k.

Rudimentary kmer counting, very unoptimized

KmerGMA.kmerFreq — Function

kmerFreq(k::Int64,
     seq::LongSequence{DNAAlphabet{4}},
     KD::Dict{LongSequence{DNAAlphabet{4}};
     returnDict::Bool = false)

Computes the kmer frequency vector of a sequence from scratch by iteration (O(n)). It is actually a slightly slower version that the main GMA does not use.

KD is the kmer Dictionary and an optional argument. If left empty, will be automatically generated WITH N's.

Note that it includes overlap(it can be tested whether it helps. but intuitively the difference shouldnt be too large).

Other utilities

KmerGMA.avgRecLen — Function

avgRecLen(reader;
          rnd::Bool = true)

get the rounded average length of every record in a FASTA.Reader object. (it can also be a string indicating the path)

rnd is an optional argument for whether the result should be rounded to an Interger.

KmerGMA.kfv — Function

kfv(maindict::Dict{LongSequence{DNAAlphabet{4}}, Float64},
    KD::Dict{LongSequence{DNAAlphabet{4}}, Int64})

Converts a dictionary of kmers and COUNTS (maindict) to a kmer frequency VECTOR.

KD is the kmer dictionary and must be based on the the same kmer length as the count dictionary.

KmerGMA.percentN — Function

percentN(seq)

Simple function to count the percentage of N nucleotides in a longsequence, record, or reader object.

KmerGMA.readerNTs — Function

readerNTs(reader::FASTX.FASTA.Reader)

function to count the number of nts in a fasta reader object (length)

KmerGMA.readerlens — Function

readerlens(reader::FASTX.FASTA.Reader)

function to just look at the individual record lengths of a reader object and returns a vector of lengths

KmerGMA.flipDict — Function

flipDict(dict::Dict{LongSequence{DNAAlphabet{4}},Int64})

function to return a NEW dictionary that has the key value pairs flipped compared to the input.