Utility functions for processing sequences and kmers

fasta file utilities

KmerGMA.fasta_id_to_cumulative_len_dict — Function

fasta_id_to_cumulative_len_dict(fasta_file_path::String)

function to record and store cumulative lengths of the BEGINNING of each record in a dictionary, from a fasta file. for example: for the following fasta file

>firstseq
ATGC
>secondseq
AT

The dictionary returned would be:

"firstseq"  => 4
"secondseq" => 6

Not nessecarily in that order.

KmerGMA.getSeq — Function

getSeq(seq::FASTA.Record)

Alias to get the dna longsequence of a fasta record

kmer utilities

KmerGMA.kmer_count — Function

kmer_count(str::DnaSeq, k::Int, Nt_bits::DnaBits = NUCLEOTIDE_BITS)

Simple kmer counting function for a DNA BioSequence str, where k is the kmer length to count. The function returns a kmer frequency vector where each INDEX of the vector corresponds to a unique kmer. For each index, the unsigned bits of the index correspond to a binary representation of sequences where two bits encode one nucleotide. To see what kmer each index corresponds to, see the function as_kmer

KmerGMA.as_kmer — Function

get_kmer(kmer_uint::UInt, kmer_len::Int,Nt_bits::Dict{UInt, Seq} = BitNtDict)

Takes a unsigned integer representing a 2-bit-based dna kmer, and the actual length of the kmer, and returns the BioSequences dna sequence of the kmer. The Nt_bits argument can be ignored. Additionally, note that both input parameters will be modified to 0 or 1.

KmerGMA.as_UInt — Function

UInt(kmer_seq::LongSequence{DNAAlphabet{4}}, Nt_bits::DnaBits = NUCLEOTIDE_BITS)

Convert a dna biosequence kmer_seq into an unsigned integer. Users can ignore the second argument