Utilities

These are a collection of functions to do with counting sizes, record sizes, and dictionary manipulation functions.

note!!! They are just here as this is a preliminary version of the package.

KmerGMA.avgRecLen — Function

avgRecLen(reader;
          rnd::Bool = true)

get the rounded average length of every record in a FASTA.Reader object. (it can also be a string indicating the path)

rnd is an optional argument for whether the result should be rounded to an Interger.

KmerGMA.kfv — Function

kfv(maindict::Dict{LongSequence{DNAAlphabet{4}}, Float64},
    KD::Dict{LongSequence{DNAAlphabet{4}}, Int64})

Converts a dictionary of kmers and COUNTS (maindict) to a kmer frequency VECTOR.

KD is the kmer dictionary and must be based on the the same kmer length as the count dictionary.

KmerGMA.percentN — Function

percentN(seq)

Simple function to count the percentage of N nucleotides in a longsequence, record, or reader object.

KmerGMA.readerNTs — Function

readerNTs(reader::FASTX.FASTA.Reader)

function to count the number of nts in a fasta reader object (length)

KmerGMA.readerlens — Function

readerlens(reader::FASTX.FASTA.Reader)

function to just look at the individual record lengths of a reader object and returns a vector of lengths